Versión 4

+++ Login/src/main/java/modelo/detalleventasDAO.java
@@ -24,8 +24,8 @@ public class detalleventasDAO {
 			
 			ps =cnn.prepareStatement("Insert Into detalle_venta(cod_venta,cant_producto,cod_producto,valor_venta,valor_iva, valor_total)   value(?,?,?,?,?,?);");
 			ps.setInt(1, det.getCod_venta());
-			ps.setLong(2, det.getCant_producto());
-			ps.setLong(3, det.getCod_producto());
+			ps.setInt(2, det.getCant_producto());
+			ps.setInt(3, det.getCod_producto());
 			ps.setDouble(4, det.getValor_venta());
 			ps.setDouble(5, det.getValor_iva());
 		ps.setDouble(6, det.getValor_total());
